fix left extra uint conversion

+++ bridge-history-api/cross_msg/batch_info_fetcher.go
@@ -86,7 +86,7 @@ func (b *BatchInfoFetcher) fetchBatchInfo() error {
 	} else {
 		startHeight = latestBatch.CommitHeight + 1
 	}
-	for from := startHeight; number >= from+uint64(b.confirmation); from += uint64(fetchLimit) {
+	for from := startHeight; number >= from+b.confirmation; from += uint64(fetchLimit) {
 		to := from + uint64(fetchLimit) - 1
 		// number - confirmation can never less than 0 since the for loop condition
 		// but watch out the overflow
